build: Exceptions around Account and Transaction

+++ ms-accounts/application/usecase/src/main/java/ec/com/banking/usecase/account/DeleteAccountUseCase.java
@@ -1,5 +1,6 @@
 package ec.com.banking.usecase.account;
 
+import ec.com.banking.core.exception.AccountNotFoundException;
 import ec.com.banking.core.gateway.AccountRepository;
 
 public class DeleteAccountUseCase {
@@ -10,6 +11,9 @@ public class DeleteAccountUseCase {
     }
 
     public void execute(Long id) {
+        if(accountRepository.findById(id) == null) {
+            throw new AccountNotFoundException(id);
+        }
         accountRepository.delete(id);
     }
 }
